PUB-2234 - Updated base class
