+++ src/test/java/uk/gov/hmcts/reform/pip/publication/services/client/SpringBootTestBase.java
@@ -1,0 +1,21 @@
+package uk.gov.hmcts.reform.pip.publication.services.client;
+
+import com.redis.testcontainers.RedisContainer;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+import org.testcontainers.utility.DockerImageName;
+
+@Testcontainers
+public class SpringBootTestBase {
+
+    @Container
+    private static RedisContainer redisContainer = new RedisContainer(DockerImageName.parse("redis:latest")).withExposedPorts(6379);
+
+    @DynamicPropertySource
+    static void registerRedisProperties(DynamicPropertyRegistry registry) {
+        registry.add("spring.data.redis.host", redisContainer::getHost);
+        registry.add("spring.data.redis.port", () -> redisContainer.getMappedPort(6379).toString());
+    }
+}
